+++ frontend/src/views/sql-editor/EditorCommon/ResultView/DataTable/DataTable/index.ts
@@ -1,0 +1,3 @@
+import DataTable from "./DataTable.vue";
+
+export default DataTable;
