+++ frontend/src/views/sql-editor/EditorCommon/ResultView/DataTable/DataTableLite/TableCell.vue
@@ -1,6 +1,10 @@
 <!-- eslint-disable vue/no-v-html -->
 <template>
-  <div class="relative px-2 py-1" :class="classes" @click="handleClick">
+  <div
+    class="relative px-2 py-1 text-sm dark:text-gray-100 leading-5 whitespace-nowrap break-all"
+    :class="classes"
+    @click="handleClick"
+  >
     <div
       ref="wrapperRef"
       class="whitespace-nowrap font-mono text-start line-clamp-1"
@@ -23,19 +27,20 @@
 
 <script setup lang="ts">
 import { type Table } from "@tanstack/vue-table";
-import { useResizeObserver } from "@vueuse/core";
 import { escape } from "lodash-es";
 import { NButton } from "naive-ui";
+import stringWidth from "string-width";
 import { computed, ref } from "vue";
 import { useConnectionOfCurrentSQLEditorTab } from "@/store";
 import { Engine } from "@/types/proto/v1/common";
 import type { QueryRow, RowValue } from "@/types/proto/v1/sql_service";
 import { extractSQLRowValue, getHighlightHTMLByRegExp } from "@/utils";
-import { useSQLResultViewContext } from "../context";
+import { useSQLResultViewContext } from "../../context";
 
 const props = defineProps<{
   table: Table<QueryRow>;
   value: RowValue;
+  width: number;
   setIndex: number;
   rowIndex: number;
   colIndex: number;
@@ -44,18 +49,19 @@ const props = defineProps<{
 const { dark, disallowCopyingData, detail, keyword } =
   useSQLResultViewContext();
 const wrapperRef = ref<HTMLDivElement>();
-const truncated = ref(false);
-
-useResizeObserver(wrapperRef, (entries) => {
-  const div = entries[0].target as HTMLDivElement;
-  const contentWidth = div.scrollWidth;
-  const visibleWidth = div.offsetWidth;
-  if (contentWidth > visibleWidth) {
-    truncated.value = true;
-  } else {
-    truncated.value = false;
-  }
+const plainValue = computed(() => {
+  return extractSQLRowValue(props.value).plain;
 });
+const truncated = computed(() => {
+  // not that accurate
+  const content = String(plainValue.value);
+  const em = 8;
+  const padding = 8;
+  const guessedContentWidth = stringWidth(content) * em + padding * 2;
+
+  return guessedContentWidth > props.width;
+});
+
 const { database } = useConnectionOfCurrentSQLEditorTab();
 
 const clickable = computed(() => {
@@ -84,7 +90,7 @@ const classes = computed(() => {
 });
 
 const html = computed(() => {
-  const value = extractSQLRowValue(props.value).plain;
+  const value = plainValue.value;
   if (value === undefined) {
     return `<span class="text-gray-400 italic">UNSET</span>`;
   }
