+++ frontend/src/views/sql-editor/EditorCommon/ResultView/DataTable/common/SensitiveDataIcon.vue
